Move inner test case classes to upper level.
