+++ registration/src/test/java/org/spine3/samples/lobby/registration/seat/availability/SeatsAvailabilityAggregateShould.java
@@ -20,30 +20,23 @@
 
 package org.spine3.samples.lobby.registration.seat.availability;
 
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
 import org.junit.Test;
 import org.spine3.base.CommandContext;
-import org.spine3.samples.lobby.common.ConferenceId;
-import org.spine3.samples.lobby.common.ReservationId;
-import org.spine3.samples.lobby.common.SeatTypeId;
 import org.spine3.samples.lobby.registration.contracts.SeatQuantity;
-import org.spine3.samples.lobby.registration.seat.availability.testcase.TestCase;
+import org.spine3.samples.lobby.registration.seat.availability.testcase.*;
 
 import java.util.List;
 import java.util.Map;
 
-import static org.junit.Assert.*;
-import static org.spine3.samples.lobby.common.util.IdFactory.*;
-import static org.spine3.samples.lobby.registration.util.Seats.*;
-import static org.spine3.util.Identifiers.newUuid;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
+import static org.spine3.samples.lobby.registration.util.Seats.newSeatQuantity;
 
 /**
  * @author Alexander Litus
  */
-@SuppressWarnings({"InstanceMethodNamingConvention", "TypeMayBeWeakened", "UtilityClass",
-        "MagicNumber", "ClassWithTooManyMethods", "UtilityClassWithoutPrivateConstructor", "RefusedBequest", "EmptyClass",
-        "AbstractClassWithoutAbstractMethods", "NoopMethodInAbstractClass", "LocalVariableNamingConvention"})
+@SuppressWarnings({"InstanceMethodNamingConvention", "MagicNumber", "ClassWithTooManyMethods",
+        "LocalVariableNamingConvention", "OverlyCoupledClass"})
 public class SeatsAvailabilityAggregateShould {
 
     private final TestSeatsAvailabilityAggregate defaultAggregate = new TestCase().givenAggregate();
@@ -389,510 +382,5 @@ public class SeatsAvailabilityAggregateShould {
         final SeatQuantity primaryQuantity = testCase.getPrimarySeat();
         final SeatQuantity expectedResult = newSeatQuantity(primaryQuantity.getSeatTypeId(), expectedNewQuantity);
         assertTrue(availableSeatsUpdated.contains(expectedResult));
-    }
-
-    private abstract static class MakeSeatReservationCmdHandling extends TestCase {
-
-        protected static final ConferenceId CONFERENCE_ID = newConferenceId();
-        protected static final ReservationId RESERVATION_ID = newReservationId();
-
-        protected static final SeatTypeId MAIN_SEAT_TYPE_ID = newSeatTypeId("main-" + newUuid());
-        protected static final SeatTypeId WORKSHOP_SEAT_TYPE_ID = newSeatTypeId("workshop-" + newUuid());
-
-        private static final int MAIN_SEAT_COUNT_AVAILABLE = 100;
-        private static final int WORKSHOP_SEAT_COUNT_AVAILABLE = 70;
-
-        private static final List<SeatQuantity> AVAILABLE_SEATS = ImmutableList.of(
-                newSeatQuantity(MAIN_SEAT_TYPE_ID, MAIN_SEAT_COUNT_AVAILABLE),
-                newSeatQuantity(WORKSHOP_SEAT_TYPE_ID, WORKSHOP_SEAT_COUNT_AVAILABLE));
-
-        protected List<SeatQuantity> getAvailableSeats() {
-            //noinspection ReturnOfCollectionOrArrayField
-            return AVAILABLE_SEATS;
-        }
-
-        protected static class Command {
-
-            static final int MAIN_SEAT_COUNT_REQUESTED = 10;
-            static final int WORKSHOP_SEAT_COUNT_REQUESTED = 7;
-
-            static final MakeSeatReservation MAKE_SEAT_RESERVATION = MakeSeatReservation.newBuilder()
-                    .setConferenceId(CONFERENCE_ID)
-                    .setReservationId(RESERVATION_ID)
-                    .addSeat(newSeatQuantity(MAIN_SEAT_TYPE_ID, MAIN_SEAT_COUNT_REQUESTED))
-                    .addSeat(newSeatQuantity(WORKSHOP_SEAT_TYPE_ID, WORKSHOP_SEAT_COUNT_REQUESTED))
-                    .build();
-        }
-        
-        protected MakeSeatReservation givenCommand() {
-            return Command.MAKE_SEAT_RESERVATION;
-        }
-        
-        protected CommandContext givenCommandContext() {
-            return CommandContext.getDefaultInstance();
-        }
-
-        protected void validateResult(SeatsReserved event, MakeSeatReservation cmd) {
-            assertEquals(cmd.getConferenceId(), event.getConferenceId());
-            assertEquals(cmd.getReservationId(), event.getReservationId());
-
-            final List<SeatQuantity> requestedSeats = cmd.getSeatList();
-            final List<SeatQuantity> reservedSeats = event.getReservedSeatUpdatedList();
-            checkReservedSeatsUpdated(reservedSeats, requestedSeats);
-
-            final List<SeatQuantity> availableSeatsUpdated = event.getAvailableSeatUpdatedList();
-            checkAvailableSeatsUpdated(availableSeatsUpdated, requestedSeats);
-        }
-
-        protected abstract void checkReservedSeatsUpdated(List<SeatQuantity> reservedSeats, List<SeatQuantity> requestedSeats);
-
-        protected void checkAvailableSeatsUpdated(List<SeatQuantity> availableSeats, List<SeatQuantity> requestedSeats) {
-            assertEquals(requestedSeats.size(), availableSeats.size());
-
-            final int expectedMainSeatCount = MAIN_SEAT_COUNT_AVAILABLE - Command.MAIN_SEAT_COUNT_REQUESTED;
-            final SeatQuantity mainSeatAvailable = findById(availableSeats, MAIN_SEAT_TYPE_ID);
-            assertEquals(expectedMainSeatCount, mainSeatAvailable.getQuantity());
-
-            final int expectedWorkshopSeatCount = WORKSHOP_SEAT_COUNT_AVAILABLE - Command.WORKSHOP_SEAT_COUNT_REQUESTED;
-            final SeatQuantity workshopSeatAvailable = findById(availableSeats, WORKSHOP_SEAT_TYPE_ID);
-            assertEquals(expectedWorkshopSeatCount, workshopSeatAvailable.getQuantity());
-        }
-
-        private static SeatTypeId newSeatTypeId(String uuid) {
-            return SeatTypeId.newBuilder().setUuid(uuid).build();
-        }
-    }
-
-    private static class EnoughSeatsAndNoPendingReservations extends MakeSeatReservationCmdHandling {
-
-        @Override
-        public TestSeatsAvailabilityAggregate givenAggregate() {
-            final TestSeatsAvailabilityAggregate aggregate = super.givenAggregate();
-            final SeatsAvailability state = aggregate.getState().toBuilder()
-                    .addAllAvailableSeat(getAvailableSeats())
-                    .build();
-            aggregate.incrementState(state);
-            return aggregate;
-        }
-
-        @Override
-        protected void checkReservedSeatsUpdated(List<SeatQuantity> reservedSeats, List<SeatQuantity> requestedSeats) {
-            assertEquals(requestedSeats.size(), reservedSeats.size());
-            assertTrue(requestedSeats.containsAll(reservedSeats));
-        }
-    }
-
-    private static class EnoughSeatsAndExistPendingReservations extends MakeSeatReservationCmdHandling {
-
-        private static final int MAIN_SEATS_PENDING_COUNT = 5;
-
-        private static final int EXPECTED_RESERVED_MAIN_SEAT_COUNT =
-                Command.MAIN_SEAT_COUNT_REQUESTED + MAIN_SEATS_PENDING_COUNT;
-
-        private static final ImmutableMap<String, SeatQuantities> PENDING_RESERVATIONS =
-                ImmutableMap.<String, SeatQuantities>builder()
-                        .put(
-                                RESERVATION_ID.getUuid(),
-                                newSeatQuantities(newSeatQuantity(MAIN_SEAT_TYPE_ID, MAIN_SEATS_PENDING_COUNT))
-                        ).build();
-
-        @Override
-        public TestSeatsAvailabilityAggregate givenAggregate() {
-            final TestSeatsAvailabilityAggregate aggregate = super.givenAggregate();
-            final SeatsAvailability state = aggregate.getState().toBuilder()
-                    .addAllAvailableSeat(getAvailableSeats())
-                    .putAllPendingReservations(PENDING_RESERVATIONS)
-                    .build();
-            aggregate.incrementState(state);
-            return aggregate;
-        }
-
-        @Override
-        protected void checkReservedSeatsUpdated(List<SeatQuantity> reservedSeats, List<SeatQuantity> requestedSeats) {
-            assertEquals(requestedSeats.size(), reservedSeats.size());
-
-            final SeatQuantity mainSeatReserved = findById(reservedSeats, MAIN_SEAT_TYPE_ID);
-            assertEquals(EXPECTED_RESERVED_MAIN_SEAT_COUNT, mainSeatReserved.getQuantity());
-
-            final SeatQuantity workshopSeatReserved = findById(reservedSeats, WORKSHOP_SEAT_TYPE_ID);
-            assertTrue(requestedSeats.contains(workshopSeatReserved));
-        }
-    }
-
-    private static class NotEnoughSeatsAndNoPendingReservations extends MakeSeatReservationCmdHandling {
-
-        private static final int MAIN_SEAT_COUNT_AVAILABLE = 5;
-
-        private static final int WORKSHOP_SEAT_COUNT_AVAILABLE = 30;
-        private static final int EXPECTED_RESERVED_WORKSHOP_SEAT_COUNT =
-                WORKSHOP_SEAT_COUNT_AVAILABLE - Command.WORKSHOP_SEAT_COUNT_REQUESTED;
-
-        private static final List<SeatQuantity> AVAILABLE_SEATS = ImmutableList.of(
-                newSeatQuantity(MAIN_SEAT_TYPE_ID, MAIN_SEAT_COUNT_AVAILABLE),
-                newSeatQuantity(WORKSHOP_SEAT_TYPE_ID, WORKSHOP_SEAT_COUNT_AVAILABLE));
-
-        @Override
-        public TestSeatsAvailabilityAggregate givenAggregate() {
-            final TestSeatsAvailabilityAggregate aggregate = super.givenAggregate();
-            final SeatsAvailability state = aggregate.getState().toBuilder()
-                    .addAllAvailableSeat(AVAILABLE_SEATS)
-                    .build();
-            aggregate.incrementState(state);
-            return aggregate;
-        }
-
-        @Override
-        protected void checkReservedSeatsUpdated(List<SeatQuantity> reservedSeats, List<SeatQuantity> requestedSeats) {
-            assertEquals(requestedSeats.size(), reservedSeats.size());
-
-            final SeatQuantity mainSeatReserved = findById(reservedSeats, MAIN_SEAT_TYPE_ID);
-            assertEquals(MAIN_SEAT_COUNT_AVAILABLE, mainSeatReserved.getQuantity());
-
-            final SeatQuantity workshopSeatReserved = findById(reservedSeats, WORKSHOP_SEAT_TYPE_ID);
-            assertTrue(requestedSeats.contains(workshopSeatReserved));
-        }
-
-        @Override
-        protected void checkAvailableSeatsUpdated(List<SeatQuantity> availableSeats, List<SeatQuantity> requestedSeats) {
-            assertEquals(requestedSeats.size(), availableSeats.size());
-
-            final SeatQuantity mainSeatAvailable = findById(availableSeats, MAIN_SEAT_TYPE_ID);
-            assertEquals(0, mainSeatAvailable.getQuantity());
-
-            final SeatQuantity workshopSeatAvailable = findById(availableSeats, WORKSHOP_SEAT_TYPE_ID);
-            assertEquals(EXPECTED_RESERVED_WORKSHOP_SEAT_COUNT, workshopSeatAvailable.getQuantity());
-        }
-    }
-
-    private abstract static class CommitSeatReservationCmdHandling extends TestCase {
-
-        private static final ReservationId RESERVATION_ID = newReservationId();
-
-        private static final CommitSeatReservation COMMIT_SEAT_RESERVATION = CommitSeatReservation.newBuilder()
-                .setReservationId(RESERVATION_ID)
-                .build();
-
-        protected CommitSeatReservation givenCommand() {
-            return COMMIT_SEAT_RESERVATION;
-        }
-
-        protected CommandContext givenCommandContext() {
-            return CommandContext.getDefaultInstance();
-        }
-
-        protected void validateResult(SeatsReservationCommitted event, CommitSeatReservation cmd) {
-            assertEquals(cmd.getReservationId(), event.getReservationId());
-        }
-
-        private static class ExistsPendingReservation extends CommitSeatReservationCmdHandling {
-
-            private static final ImmutableMap<String, SeatQuantities> PENDING_RESERVATIONS =
-                    ImmutableMap.<String, SeatQuantities>builder()
-                            .put(RESERVATION_ID.getUuid(), newSeatQuantities(newSeatQuantity(20))).build();
-
-            @Override
-            public TestSeatsAvailabilityAggregate givenAggregate() {
-                final TestSeatsAvailabilityAggregate aggregate = super.givenAggregate();
-                final SeatsAvailability state = aggregate.getState().toBuilder()
-                        .putAllPendingReservations(PENDING_RESERVATIONS)
-                        .build();
-                aggregate.incrementState(state);
-                return aggregate;
-            }
-        }
-
-        static class EmptyState extends CommitSeatReservationCmdHandling {
-        }
-    }
-
-    private abstract static class CancelSeatReservationCmdHandling extends TestCase {
-
-        private static final ReservationId RESERVATION_ID = newReservationId();
-
-        private static final CancelSeatReservation CANCEL_SEAT_RESERVATION = CancelSeatReservation.newBuilder()
-                .setReservationId(RESERVATION_ID)
-                .setConferenceId(newConferenceId())
-                .build();
-
-        protected CancelSeatReservation givenCommand() {
-            return CANCEL_SEAT_RESERVATION;
-        }
-
-        protected CommandContext givenCommandContext() {
-            return CommandContext.getDefaultInstance();
-        }
-
-        protected void validateResult(SeatsReservationCancelled event, CancelSeatReservation cmd) {
-        }
-
-        private static class ExistsPendingReservation extends CancelSeatReservationCmdHandling {
-
-            private static final SeatQuantities TMP_RESERVED_SEATS = newSeatQuantities(
-                    newSeatQuantity(10), newSeatQuantity(20));
-
-            private static final ImmutableMap<String, SeatQuantities> PENDING_RESERVATIONS =
-                    ImmutableMap.<String, SeatQuantities>builder()
-                            .put(RESERVATION_ID.getUuid(), TMP_RESERVED_SEATS).build();
-
-            @Override
-            public TestSeatsAvailabilityAggregate givenAggregate() {
-                final TestSeatsAvailabilityAggregate aggregate = super.givenAggregate();
-                final SeatsAvailability state = aggregate.getState().toBuilder()
-                        .putAllPendingReservations(PENDING_RESERVATIONS)
-                        .build();
-                aggregate.incrementState(state);
-                return aggregate;
-            }
-
-            @Override
-            protected void validateResult(SeatsReservationCancelled event, CancelSeatReservation cmd) {
-                assertEquals(cmd.getReservationId(), event.getReservationId());
-                assertEquals(cmd.getConferenceId(), event.getConferenceId());
-
-                final List<SeatQuantity> availableSeatsActual = event.getAvailableSeatUpdatedList();
-                assertEquals(TMP_RESERVED_SEATS.getItemList(), availableSeatsActual);
-            }
-        }
-
-        static class EmptyState extends CancelSeatReservationCmdHandling {
-        }
-    }
-
-    private static class AddSeatsCmdHandling extends TestCase {
-
-        private static final AddSeats ADD_SEATS = AddSeats.newBuilder()
-                .setQuantity(newSeatQuantity(5))
-                .setConferenceId(newConferenceId())
-                .build();
-
-        protected AddSeats givenCommand() {
-            return ADD_SEATS;
-        }
-
-        protected CommandContext givenCommandContext() {
-            return CommandContext.getDefaultInstance();
-        }
-
-        @Override
-        public TestSeatsAvailabilityAggregate givenAggregate() {
-            final TestSeatsAvailabilityAggregate aggregate = super.givenAggregate();
-            final SeatsAvailability state = aggregate.getState().toBuilder()
-                    .addAvailableSeat(newSeatQuantity(20))
-                    .build();
-            aggregate.incrementState(state);
-            return aggregate;
-        }
-
-        protected void validateResult(AddedAvailableSeats event, AddSeats cmd) {
-            final SeatQuantity quantityToAdd = cmd.getQuantity();
-            final SeatQuantity quantityAdded = event.getQuantity();
-            assertEquals(quantityToAdd, quantityAdded);
-        }
-    }
-
-    private static class RemoveSeatsCmdHandling extends TestCase {
-
-        private static final SeatTypeId SEAT_TYPE_ID = newSeatTypeId();
-        private static final int SEAT_QUANTITY_TO_REMOVE = 20;
-
-        private static final SeatQuantity AVAILABLE_QUANTITY = newSeatQuantity(SEAT_TYPE_ID, 80);
-
-        private static final RemoveSeats REMOVE_SEATS = RemoveSeats.newBuilder()
-                .setQuantity(newSeatQuantity(SEAT_TYPE_ID, SEAT_QUANTITY_TO_REMOVE))
-                .setConferenceId(newConferenceId())
-                .build();
-
-        protected RemoveSeats givenCommand() {
-            return REMOVE_SEATS;
-        }
-
-        protected CommandContext givenCommandContext() {
-            return CommandContext.getDefaultInstance();
-        }
-
-        @Override
-        public TestSeatsAvailabilityAggregate givenAggregate() {
-            final TestSeatsAvailabilityAggregate aggregate = super.givenAggregate();
-            final SeatsAvailability state = aggregate.getState().toBuilder()
-                    .addAvailableSeat(AVAILABLE_QUANTITY)
-                    .build();
-            aggregate.incrementState(state);
-            return aggregate;
-        }
-
-        protected void validateResult(RemovedAvailableSeats event, RemoveSeats cmd) {
-            assertEquals(cmd.getQuantity(), event.getQuantity());
-        }
-
-        static class EmptyState extends RemoveSeatsCmdHandling {
-            @Override
-            public TestSeatsAvailabilityAggregate givenAggregate() {
-                return new TestSeatsAvailabilityAggregate(newSeatsAvailabilityId());
-            }
-        }
-    }
-
-    private static class ExistAvailableSeatsAndPendingReservations extends TestCase {
-
-        private static final ReservationId RESERVATION_ID = newReservationId();
-
-        private static final SeatQuantities RESERVED_SEATS = newSeatQuantities(
-                newSeatQuantity(56), newSeatQuantity(38));
-
-        private static final ImmutableMap<String, SeatQuantities> PENDING_RESERVATIONS = ImmutableMap.<String, SeatQuantities>builder()
-                .put(RESERVATION_ID.getUuid(), RESERVED_SEATS).build();
-
-        private static final List<SeatQuantity> AVAILABLE_SEATS = ImmutableList.of(
-                newSeatQuantity(110),
-                newSeatQuantity(220));
-
-        @Override
-        public TestSeatsAvailabilityAggregate givenAggregate() {
-            final TestSeatsAvailabilityAggregate aggregate = super.givenAggregate();
-            final SeatsAvailability state = aggregate.getState().toBuilder()
-                    .addAllAvailableSeat(AVAILABLE_SEATS)
-                    .putAllPendingReservations(PENDING_RESERVATIONS)
-                    .build();
-            aggregate.incrementState(state);
-            return aggregate;
-        }
-
-        protected static List<SeatQuantity> getAvailableSeats() {
-            //noinspection ReturnOfCollectionOrArrayField
-            return AVAILABLE_SEATS;
-        }
-
-        protected static SeatQuantities getReservedSeats() {
-            return RESERVED_SEATS;
-        }
-
-        protected static ReservationId getReservationId() {
-            return RESERVATION_ID;
-        }
-    }
-
-    private static class SeatsReservedEventApplying extends ExistAvailableSeatsAndPendingReservations {
-
-        private static final SeatsReserved SEATS_RESERVED = SeatsReserved.newBuilder()
-                .setConferenceId(newConferenceId())
-                .setReservationId(getReservationId())
-                .addReservedSeatUpdated(newSeatQuantity(256))
-                .addAvailableSeatUpdated(newSeatQuantity(24)).build();
-
-        static {
-            assertNotEquals(getAvailableSeats(), SEATS_RESERVED.getAvailableSeatUpdatedList());
-            assertNotEquals(getReservedSeats(), SEATS_RESERVED.getReservedSeatUpdatedList());
-        }
-
-        public SeatsReserved givenEvent() {
-            return SEATS_RESERVED;
-        }
-    }
-
-    private static class SeatsReservationCommittedEventApplying extends ExistAvailableSeatsAndPendingReservations {
-
-        private static final SeatsReservationCommitted SEATS_RESERVATION_COMMITTED = SeatsReservationCommitted.newBuilder()
-                .setReservationId(getReservationId()).build();
-
-        public SeatsReservationCommitted givenEvent() {
-            return SEATS_RESERVATION_COMMITTED;
-        }
-    }
-
-    private static class SeatsReservationCancelledEventApplying extends ExistAvailableSeatsAndPendingReservations {
-
-        private static final SeatsReservationCancelled SEATS_RESERVATION_CANCELLED = SeatsReservationCancelled.newBuilder()
-                .setReservationId(getReservationId())
-                .setConferenceId(newConferenceId())
-                .addAvailableSeatUpdated(newSeatQuantity(32))
-                .build();
-
-        static {
-            assertNotEquals(getAvailableSeats(), SEATS_RESERVATION_CANCELLED.getAvailableSeatUpdatedList());
-        }
-
-        public SeatsReservationCancelled givenEvent() {
-            return SEATS_RESERVATION_CANCELLED;
-        }
-    }
-
-    private abstract static class AddedAvailableSeatsEventApplying extends ExistAvailableSeatsAndPendingReservations {
-
-        private static final SeatQuantity PRIMARY_SEAT = getAvailableSeats().get(0);
-
-        protected SeatQuantity getPrimarySeat() {
-            return PRIMARY_SEAT;
-        }
-
-        protected AddedAvailableSeats givenEvent() {
-            final AddedAvailableSeats event = AddedAvailableSeats.newBuilder()
-                    .setQuantity(getQuantityToAdd())
-                    .build();
-            return event;
-        }
-
-        protected abstract SeatQuantity getQuantityToAdd();
-
-        private static class AddingNewSeatType extends AddedAvailableSeatsEventApplying {
-
-            private final SeatQuantity quantityToAdd = newSeatQuantity(12);
-
-            protected AddingNewSeatType() {
-                assertFalse(getAvailableSeats().contains(quantityToAdd));
-            }
-
-            @Override
-            protected SeatQuantity getQuantityToAdd() {
-                return quantityToAdd;
-            }
-        }
-
-        private static class UpdatingSeatType extends AddedAvailableSeatsEventApplying {
-
-            private final SeatQuantity quantityToAdd = newSeatQuantity(getPrimarySeat().getSeatTypeId(), 10);
-
-            @Override
-            protected SeatQuantity getQuantityToAdd() {
-                return quantityToAdd;
-            }
-        }
-    }
-
-    private abstract static class RemovedAvailableSeatsEventApplying extends ExistAvailableSeatsAndPendingReservations {
-
-        private SeatQuantity primarySeat;
-
-        protected RemovedAvailableSeats givenEvent() {
-            primarySeat = getAvailableSeats().get(0);
-            final int quantityToRemove = getQuantityToRemove();
-            final RemovedAvailableSeats event = RemovedAvailableSeats.newBuilder()
-                    .setQuantity(newSeatQuantity(primarySeat.getSeatTypeId(), quantityToRemove))
-                    .build();
-            return event;
-        }
-
-        protected SeatQuantity getPrimarySeat() {
-            return primarySeat;
-        }
-
-        protected abstract int getQuantityToRemove();
-
-        private static class RemovingLessSeatsThanRemaining extends RemovedAvailableSeatsEventApplying {
-
-            @Override
-            protected int getQuantityToRemove() {
-                final int quantityToRemove = getPrimarySeat().getQuantity() / 2;
-                return quantityToRemove;
-            }
-        }
-
-        private static class RemovingMoreSeatsThanRemaining extends RemovedAvailableSeatsEventApplying {
-
-            @Override
-            protected int getQuantityToRemove() {
-                final int quantityToRemove = getPrimarySeat().getQuantity() + 5;
-                return quantityToRemove;
-            }
-        }
     }
 }
